+++ packages/web/src/lib/posthogEvents.ts
@@ -537,6 +537,7 @@ export type PosthogEventMap = {
     },
     //////////////////////////////////////////////////////////////////
     wa_askgh_login_wall_prompted: {},
+    wa_publicsaas_cs_login_wall_prompted: {},
     //////////////////////////////////////////////////////////////////
     askgh_repo_index_requested: {
         owner: string,